requirements.txt
boto==2.38.0
dateparser==0.6.0
django==1.11.6
django-debug-toolbar==1.7
django-date-extensions==2.0
django-flat-theme==1.1.3
django-localflavor==1.6
django-robots==3.0
django-sitetree==1.8.0
django-storages-redux==1.3.1
django-taggit==0.17.6
django-tinymce
docutils==0.12
Pillow==4.1.0
pyinotify==0.9.6
pytz==2017.2
python-dateutil==2.5.3
raven==5.1.1
requests==2.18.4
six==1.10.0
sorl-thumbnail==12.3
sphinx==1.3.4
